requirements.txt
Flask>=2.0.0,<3.0.0
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.2
PyJWT==2.8.0
gunicorn==22.0.0